requirements.txt
# Core packages
numpy>=1.22.0
pandas>=1.4.0
matplotlib>=3.5.0
tqdm>=4.64.0
Pillow>=9.2.0
requests>=2.27.0

# NLP and ML packages
transformers>=4.27.0
torch>=2.0.0
accelerate>=0.20.0
einops>=0.6.0
nltk>=3.7.0
scikit-learn>=1.0.0

# OCR tools
pytesseract>=0.3.10
opencv-python>=4.6.0

# For file handling and data processing
h5py>=3.7.0
jsonlines>=3.0.0
PyYAML>=6.0.0

# For evaluations and logging
tensorboard>=2.10.0
wandb>=0.13.5  # Optional for experiment tracking

# For environment and setup
python-dotenv>=0.20.0 